backend/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
numpy>=1.24.0
tqdm>=4.65.0
flask>=2.3.0
flask-cors>=4.0.0
